requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
fastapi==0.75.1
fonttools==4.32.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
